requirements.txt
langgraph==0.2.61
langchain==0.3.15
langchain-core==0.3.31
langchain-community==0.3.14
langchain-openai==0.2.10
python-dotenv
aiohttp
fastapi
uvicorn
